requirement.txt
mysql-python-connector
matplotlib
tkinter(pre-installed with python)
hashlib
PIL (png/jpg support)